docs/samples/python/aplos_nca/requirements.txt
boto3
requests
python-dotenv # load env values
PyJWT